refactor metadata_generation.py into smaller pieces to move toward allowing future use of pipeline without inspect metadata by outside groups
