+++ tests/data/dummy/dummy_run2-summary.csv
@@ -1,3 +1,3 @@
-Sample,total_reads,mapped_reads,percentage_aligned,10_x_pc,median_coverage,mean_coverage,median_insert_size,avg_gc,sequenced_pool_component_id,search_id,coverage_gte_10_reads,fraction_acgt_bases,num_inserts_in_consensus,num_deletions_in_consensus,consensus_seq_name,assembly_method,timestamp,se_or_pe,seq_run,consensus_s3,trimmed_bam_s3,variants_s3
-SU002__DUMLIB__DUMWELL__001,4902366,4890878,99.76566417,99.99331171,16835,23714.0253,252,38.89262152,SU002__DUMLIB__DUMWELL__001,SU002,0.99999,1,0,0,Consensus_SU003_S14_L001.trimmed.sorted.pileup.consensus_threshold_0.5_quality_20,iVar version 1.3.1,2021-02-28_10-30-04,pe,PDH_83-233854622,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/SU002__DUMLIB__DUMWELL__001/SU002__DUMLIB__DUMWELL__001.trimmed.sorted.pileup.consensus.fa,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/SU002__DUMLIB__DUMWELL__001/SU002__DUMLIB__DUMWELL__001.trimmed.sorted.bam,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/SU002__DUMLIB__DUMWELL__001/SU002__DUMLIB__DUMWELL__001.trimmed.sorted.pileup.variants.tsv
-Pos_002453_S239,168,0,0,,,,,,Pos_002453_S239,Pos_002453_S239,,,,,,iVar version 1.3.1,2021-03-13_00-11-09,pe,PDH_83-233854622,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/Pos_002453_S239/Pos_002453_S239.trimmed.sorted.pileup.consensus.fa,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/Pos_002453_S239/Pos_002453_S239.trimmed.sorted.bam,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/Pos_002453_S239/Pos_002453_S239.trimmed.sorted.pileup.variants.tsv
+Sample,total_reads,mapped_reads,percentage_aligned,10_x_pc,median_coverage,mean_coverage,median_insert_size,P25_Ins_size,avg_gc,Uncapped_Reads,Sub_Map_Pct_Aligned,Pct_Q30,sequenced_pool_component_id,search_id,coverage_gte_10_reads,fraction_acgt_bases,num_inserts_in_consensus,num_deletions_in_consensus,consensus_seq_name,assembly_method,timestamp,se_or_pe,seq_run,consensus_s3,trimmed_bam_s3,variants_s3
+SU002__DUMLIB__DUMWELL__001,4902366,4890878,99.76566417,99.99331171,16835,23714.0253,252,146,38.89262152,14520642,98,95.173,SU002__DUMLIB__DUMWELL__001,SU002,0.99999,1,0,0,Consensus_SU003_S14_L001.trimmed.sorted.pileup.consensus_threshold_0.5_quality_20,iVar version 1.3.1,2021-02-28_10-30-04,pe,PDH_83-233854622,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/SU002__DUMLIB__DUMWELL__001/SU002__DUMLIB__DUMWELL__001.trimmed.sorted.pileup.consensus.fa,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/SU002__DUMLIB__DUMWELL__001/SU002__DUMLIB__DUMWELL__001.trimmed.sorted.bam,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/SU002__DUMLIB__DUMWELL__001/SU002__DUMLIB__DUMWELL__001.trimmed.sorted.pileup.variants.tsv
+Pos_002453_S239,168,0,0,,,,,,,168,0,10.2,Pos_002453_S239,Pos_002453_S239,,,,,,iVar version 1.3.1,2021-03-13_00-11-09,pe,PDH_83-233854622,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/Pos_002453_S239/Pos_002453_S239.trimmed.sorted.pileup.consensus.fa,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/Pos_002453_S239/Pos_002453_S239.trimmed.sorted.bam,s3://ucsd-other/2021-02-08-ARTIC/2021-02-08-ARTIC_results/2021-03-22_23-30-22_pe/2021-02-08-ARTIC_samples/Pos_002453_S239/Pos_002453_S239.trimmed.sorted.pileup.variants.tsv
